fix(ui): align environment picker checks with current controls
Preserve cloud retry and dispatch contracts through the current configuration cards, fixed OS display, environment names and account inventory lifecycle. Keep Connect scope checks on the live action and disabled Git explanations on the exposed tooltip.

Remove unused cloud render exports, split where-chip coverage without changing its scenarios, and use the action cursor token. Reuse the independently reviewed upstream cleanup and test split with their original authors credited below.
