+++ ui/src/e2e/new-session-page.cloud-os.e2e.test.ts
@@ -76,14 +76,14 @@ suite.define(() => {
         const linux = picker.locator('[data-value="os:linux"]');
         await picker.locator('[data-value="cloud:aws"]').hover();
         await linux.waitFor();
-        expect(await linux.isEnabled()).toBe(true);
-        expect(await linux.getAttribute("aria-pressed")).toBe("true");
+        expect(await linux.evaluate((element) => element.tagName)).toBe("SPAN");
+        expect(await linux.getAttribute("aria-pressed")).toBeNull();
+        expect(await linux.textContent()).toBe("Linux");
         for (const os of ["macos", "windows"]) {
           const option = picker.locator(`[data-value="os:${os}"]`);
           expect(await option.count()).toBe(0);
         }
         await capturePicker("02-after-unavailable-operating-systems.png");
-        await linux.click();
         await page.keyboard.press("Escape");
         await page.locator(".new-session-page__message").fill("Continue on Linux");
         await page.getByRole("button", { name: "Start session" }).click();
