+++ ui/src/e2e/new-session-page.device-dispatch.e2e.test.ts
@@ -71,8 +71,8 @@ suite.define(() => {
             ),
         )
         .toEqual([
-          "gateway",
           "auto-device",
+          "gateway",
           "device:paired-runner",
           "device:offline-runner",
           "cloud:aws",
@@ -82,7 +82,7 @@ suite.define(() => {
           .locator(".new-session-page__environment-heading")
           .allTextContents()
           .then((headings) => headings.map((heading) => heading.replace(/\s+/g, " ").trim())),
-      ).toEqual(["Local", "Your devices", "Cloud"]);
+      ).toEqual(["Your devices", "Cloud"]);
       const auto = destinations.locator('[data-value="auto-device"]');
       expect(await auto.getAttribute("aria-pressed")).toBe("false");
       expect(await destinations.getByRole("button", { name: /^aws(?: · .+)?$/ }).count()).toBe(1);
